refactor: update file structure on client side
