+++ client/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.13.5",
     "@mui/system": "^5.13.5",
@@ -23,7 +24,6 @@
     "react-router-dom": "^6.12.1"
   },
   "devDependencies": {
-    "@emotion/styled": "^11.11.0",
     "@mui/styled-engine": "^5.13.2",
     "@types/react": "^18.2.9",
     "@types/react-dom": "^18.0.11",
